+++ lib/src/json_array.dart
@@ -46,9 +46,9 @@ import 'package:json_response/src/json_array_impl.dart';
 ///   print(jsonArray.get(index: 0));
 ///
 ///   // If you don't like the structure of several nested lists,
-///   // you can use the toFlat method to make the nested structure flat.
+///   // you can use the flatten method to make the nested structure flat.
 ///   // This method returns a new flattened JsonArray.
-///   print(jsonArray.toFlat());
+///   print(jsonArray.flatten());
 /// ```
 abstract class JsonArray {
   /// Returns the new instance of [JsonArray] based on [response].
@@ -88,7 +88,7 @@ abstract class JsonArray {
   void enumerateArray(void Function(int index, JsonArray jsonArray) action);
 
   /// Returns the new instance of flattened [JsonArray] from this [JsonArray].
-  JsonArray toFlat();
+  JsonArray flatten();
 
   /// Returns true if this json array is empty, otherwise false.
   bool get isEmpty;